Add files via upload
In this part, we want to find lacalization of cell phones using the power of Wi-Fi signals. These Wi-Fi signals recieve from several routers in a building. Suppose that there are 7 routers in a building. In each point, the signal of each router has different power. This building is divided into four regions and in each region, the power of recieved signal is measured. All the measurments are in wifi-data.txt. In this file, the power of recieved signal is in each row and the last number of each row is number of region.
a. Implement Logistic Regression for data of regions 1 and 3. First, shffle the data of regions 1 and 3. Then, get 85% of data as train set and 15% of data as test set. Report the accuracy of your classification for train and test sets and report the confusion matrix of test set.
b. Implement SVM for data of region 1 and 3.